refactor: split audio backend into platform-specific implementations with compile-time selection
- Extract PortAudio backend to audio_backend_portaudio.cpp and SDL backend to audio_backend_sdl.cpp
- Add audio_backend.h with opaque AudioBackendState and factory functions (create/destroy)
- Implement platform-conditional source selection in CMake (Emscripten uses SDL, native uses PortAudio)
- Split file_dialog.cpp into file_dialog_native.cpp and file_dialog_web.cpp for platform-specific implementations
- Move
