+++ src/audio/audio_engine.cpp
@@ -1,718 +1,30 @@
 #include "audio/audio_engine.h"
+#include "audio/audio_backend.h"
 #include <cstring>
 #include <iostream>
-#include <cctype>
 #include <algorithm>
 #include <chrono>
 
-#ifdef __EMSCRIPTEN__
-#include <SDL.h>
-#else
-#ifdef _WIN32
-#include <pa_win_wasapi.h>
-#endif
-#endif
-
 namespace GuitarAmp {
+
+// =============================================================================
+// Construction / destruction
+// =============================================================================
 
 AudioEngine::AudioEngine() {
     process_buffer_.resize(MAX_BUFFER_SIZE, 0.0f);
+    backend_ = create_audio_backend();
 }
 
 AudioEngine::~AudioEngine() {
     shutdown();
+    destroy_audio_backend(backend_);
+    backend_ = nullptr;
 }
-
-#ifndef __EMSCRIPTEN__
-
-bool AudioEngine::is_usb_device_name(const std::string& name) {
-    // Convert to lowercase for matching
-    std::string lower = name;
-    std::transform(lower.begin(), lower.end(), lower.begin(),
-                   [](unsigned char c) { return std::tolower(c); });
-
-    // Common USB guitar cable / interface identifiers
-    static const char* usb_keywords[] = {
-        "usb", "guitar", "guitar link", "irig", "scarlett",
-        "behringer", "focusrite", "presonus", "steinberg",
-        "audio interface", "line 6", "rocksmith", "umc",
-        "um2", "uphoria", "podcast", "xenyx"
-    };
-
-    for (const auto* keyword : usb_keywords) {
-        if (lower.find(keyword) != std::string::npos) {
-            return true;
-        }
-    }
-    return false;
-}
-
-static int get_host_api_priority(PaHostApiTypeId type) {
-    // Platform-aware priority: pick the lowest-latency host API per OS
-#if defined(__linux__)
-    switch (type) {
-        case paJACK:          return 100;  // JACK is king on Linux
-        case paALSA:          return 70;
-        default:              return 10;
-    }
-#elif defined(_WIN32)
-    switch (type) {
-        case paASIO:          return 100;  // ASIO still best if available
-        case paWASAPI:        return 90;   // WASAPI Exclusive is excellent
-        case paDirectSound:   return 40;
-        case paMME:           return 10;
-        default:              return 20;
-    }
-#elif defined(__APPLE__)
-    switch (type) {
-        case paCoreAudio:     return 100;
-        default:              return 30;
-    }
-#else
-    (void)type;
-    return 30;
-#endif
-}
-
-static bool is_projector_or_hdmi(const std::string& name) {
-    std::string lower = name;
-    std::transform(lower.begin(), lower.end(), lower.begin(),
-                   [](unsigned char c) { return std::tolower(c); });
-    return lower.find("epson") != std::string::npos
-        || lower.find("projector") != std::string::npos
-        || lower.find("hdmi") != std::string::npos
-        || lower.find("displayport") != std::string::npos;
-}
-
-void AudioEngine::auto_detect_devices() {
-    int device_count = Pa_GetDeviceCount();
-    int num_apis = Pa_GetHostApiCount();
-
-    // Phase 1: Print all devices
-    std::cout << "\nDetected audio devices:" << std::endl;
-    for (int i = 0; i < device_count; ++i) {
-        const PaDeviceInfo* info = Pa_GetDeviceInfo(i);
-        if (!info) continue;
-        bool is_usb = is_usb_device_name(info->name);
-        const PaHostApiInfo* api = Pa_GetHostApiInfo(info->hostApi);
-        const char* api_name = api ? api->name : "Unknown";
-        if (info->maxInputChannels > 0) {
-            std::cout << "  [IN]  " << info->name
-                      << " (" << api_name << ")"
-                      << (is_usb ? " [USB]" : "") << std::endl;
-        }
-        if (info->maxOutputChannels > 0) {
-            std::cout << "  [OUT] " << info->name
-                      << " (" << api_name << ")"
-                      << (is_usb ? " [USB]" : "") << std::endl;
-        }
-    }
-
-    // Phase 2: For each host API (ranked by priority), find the best
-    // USB input + non-USB/non-projector output PAIR on the SAME API.
-    // PortAudio requires input and output to be on the same host API.
-
-    struct ApiCandidate {
-        int api_index;
-        int priority;
-        int usb_input;       // device index, -1 if none
-        int best_output;     // device index, -1 if none
-        std::string api_name;
-    };
-
-    std::vector<ApiCandidate> candidates;
-    for (int a = 0; a < num_apis; ++a) {
-        const PaHostApiInfo* api = Pa_GetHostApiInfo(a);
-        if (!api) continue;
-
-        ApiCandidate c;
-        c.api_index = a;
-        c.priority = get_host_api_priority(api->type);
-        c.usb_input = -1;
-        c.best_output = -1;
-        c.api_name = api->name;
-
-        // Scan devices belonging to this API
-        for (int d = 0; d < api->deviceCount; ++d) {
-            int dev_idx = Pa_HostApiDeviceIndexToDeviceIndex(a, d);
-            const PaDeviceInfo* info = Pa_GetDeviceInfo(dev_idx);
-            if (!info) continue;
-
-            bool is_usb = is_usb_device_name(info->name);
-
-            // Best USB input on this API
-            if (is_usb && info->maxInputChannels > 0 && c.usb_input < 0) {
-                c.usb_input = dev_idx;
-            }
-
-            // Best non-USB, non-projector output on this API
-            if (!is_usb && !is_projector_or_hdmi(info->name)
-                && info->maxOutputChannels > 0 && c.best_output < 0) {
-                c.best_output = dev_idx;
-            }
-        }
-
-        // If no good output found, accept any non-USB output (including projector)
-        if (c.best_output < 0) {
-            for (int d = 0; d < api->deviceCount; ++d) {
-                int dev_idx = Pa_HostApiDeviceIndexToDeviceIndex(a, d);
-                const PaDeviceInfo* info = Pa_GetDeviceInfo(dev_idx);
-                if (!info) continue;
-                if (!is_usb_device_name(info->name) && info->maxOutputChannels > 0) {
-                    c.best_output = dev_idx;
-                    break;
-                }
-            }
-        }
-
-        candidates.push_back(c);
-    }
-
-    // Sort by priority (highest first)
-    std::sort(candidates.begin(), candidates.end(),
-              [](const ApiCandidate& a, const ApiCandidate& b) {
-                  return a.priority > b.priority;
-              });
-
-    // Phase 3: Pick the best API that has BOTH a USB input AND an output
-    bool found_pair = false;
-    for (auto& c : candidates) {
-        if (c.usb_input >= 0 && c.best_output >= 0) {
-            input_device_ = c.usb_input;
-            output_device_ = c.best_output;
-
-            const PaDeviceInfo* in_info = Pa_GetDeviceInfo(input_device_);
-            const PaDeviceInfo* out_info = Pa_GetDeviceInfo(output_device_);
-
-            std::cout << "\n>> Auto-selected (same API: " << c.api_name << "):" << std::endl;
-            std::cout << "   INPUT:  " << in_info->name << " [USB Guitar Cable]" << std::endl;
-            std::cout << "   OUTPUT: " << out_info->name << " [Speakers]" << std::endl;
-
-            if (in_info->defaultSampleRate > 0) {
-                sample_rate_ = static_cast<int>(in_info->defaultSampleRate);
-                std::cout << "   Rate:   " << sample_rate_ << " Hz (device native)" << std::endl;
-            }
-
-            found_pair = true;
-            break;
-        }
-    }
-
-    // Fallback: no USB input found — pick the best API with any input + output
-    if (!found_pair) {
-        for (auto& c : candidates) {
-            if (c.best_output >= 0) {
-                // Find any input on this API
-                const PaHostApiInfo* api = Pa_GetHostApiInfo(c.api_index);
-                for (int d = 0; d < api->deviceCount; ++d) {
-                    int dev_idx = Pa_HostApiDeviceIndexToDeviceIndex(c.api_index, d);
-                    const PaDeviceInfo* info = Pa_GetDeviceInfo(dev_idx);
-                    if (info && info->maxInputChannels > 0) {
-                        input_device_ = dev_idx;
-                        output_device_ = c.best_output;
-
-                        std::cout << "\n>> No USB guitar cable detected." << std::endl;
-                        std::cout << "   Using " << c.api_name << " defaults:" << std::endl;
-                        std::cout << "   INPUT:  " << info->name << std::endl;
-                        std::cout << "   OUTPUT: " << Pa_GetDeviceInfo(output_device_)->name << std::endl;
-
-                        found_pair = true;
-                        break;
-                    }
-                }
-                if (found_pair) break;
-            }
-        }
-    }
-
-    // Last resort: system defaults
-    if (!found_pair) {
-        input_device_ = Pa_GetDefaultInputDevice();
-        output_device_ = Pa_GetDefaultOutputDevice();
-        std::cout << "\n>> Using system default input/output devices." << std::endl;
-    }
-}
-
-#endif // !__EMSCRIPTEN__ (PortAudio helper functions)
 
 // =============================================================================
-// Platform-specific: initialize / shutdown / start / stop / device management
+// Effect chain management (platform-agnostic)
 // =============================================================================
-
-#ifndef __EMSCRIPTEN__
-// --- Native PortAudio implementation ---
-
-bool AudioEngine::initialize() {
-    PaError err = Pa_Initialize();
-    if (err != paNoError) {
-        std::cerr << "PortAudio init failed: " << Pa_GetErrorText(err) << std::endl;
-        return false;
-    }
-    initialized_ = true;
-
-    // Auto-detect USB guitar cable for input, laptop speakers for output
-    auto_detect_devices();
-
-    return true;
-}
-
-void AudioEngine::shutdown() {
-    stop();
-    if (initialized_) {
-        Pa_Terminate();
-        initialized_ = false;
-    }
-}
-
-bool AudioEngine::start() {
-    if (!initialized_ || running_) return false;
-
-    const PaDeviceInfo* in_dev = Pa_GetDeviceInfo(input_device_);
-    const PaDeviceInfo* out_dev = Pa_GetDeviceInfo(output_device_);
-    (void)in_dev; (void)out_dev;  // used only in platform-specific blocks below
-
-    // Compute desired latency from the user's buffer size setting.
-    // Do NOT use defaultLowInputLatency — on macOS built-in mic it can be >200ms.
-    double desired_latency = static_cast<double>(buffer_size_) / sample_rate_;
-
-    PaStreamParameters input_params;
-    input_params.device = input_device_;
-    input_params.channelCount = 1;
-    input_params.sampleFormat = paFloat32;
-    input_params.suggestedLatency = desired_latency;
-    input_params.hostApiSpecificStreamInfo = nullptr;
-
-    PaStreamParameters output_params;
-    output_params.device = output_device_;
-    output_params.channelCount = 1;
-    output_params.sampleFormat = paFloat32;
-    output_params.suggestedLatency = desired_latency;
-    output_params.hostApiSpecificStreamInfo = nullptr;
-
-#ifdef _WIN32
-    // WASAPI Exclusive Mode: bypass the Windows audio mixer for minimal latency
-    PaWasapiStreamInfo wasapi_in_info = {};
-    PaWasapiStreamInfo wasapi_out_info = {};
-    if (in_dev) {
-        const PaHostApiInfo* api = Pa_GetHostApiInfo(in_dev->hostApi);
-        if (api && api->type == paWASAPI) {
-            wasapi_in_info.size = sizeof(PaWasapiStreamInfo);
-            wasapi_in_info.hostApiType = paWASAPI;
-            wasapi_in_info.version = 1;
-            wasapi_in_info.flags = paWinWasapiExclusive;
-            input_params.hostApiSpecificStreamInfo = &wasapi_in_info;
-
-            wasapi_out_info.size = sizeof(PaWasapiStreamInfo);
-            wasapi_out_info.hostApiType = paWASAPI;
-            wasapi_out_info.version = 1;
-            wasapi_out_info.flags = paWinWasapiExclusive;
-            output_params.hostApiSpecificStreamInfo = &wasapi_out_info;
-
-            std::cout << "  Using WASAPI Exclusive Mode" << std::endl;
-        }
-    }
-#endif
-
-    // Use the user's configured buffer size for minimum latency
-    unsigned long frames = static_cast<unsigned long>(buffer_size_);
-
-    PaError err = Pa_OpenStream(
-        &stream_,
-        &input_params,
-        &output_params,
-        sample_rate_,
-        frames,
-        paClipOff | paDitherOff,
-        audio_callback,
-        this
-    );
-
-    if (err != paNoError) {
-        std::cerr << "Failed to open stream: " << Pa_GetErrorText(err) << std::endl;
-        // Fall back: try with explicit buffer size
-        std::cerr << "Retrying with buffer size " << buffer_size_ << "..." << std::endl;
-        err = Pa_OpenStream(
-            &stream_,
-            &input_params,
-            &output_params,
-            sample_rate_,
-            buffer_size_,
-            paClipOff | paDitherOff,
-            audio_callback,
-            this
-        );
-        if (err != paNoError) {
-            std::cerr << "Failed to open stream: " << Pa_GetErrorText(err) << std::endl;
-            return false;
-        }
-    }
-
-    err = Pa_StartStream(stream_);
-    if (err != paNoError) {
-        std::cerr << "Failed to start stream: " << Pa_GetErrorText(err) << std::endl;
-        Pa_CloseStream(stream_);
-        stream_ = nullptr;
-        return false;
-    }
-
-    running_ = true;
-    const PaStreamInfo* si = Pa_GetStreamInfo(stream_);
-    const PaDeviceInfo* in_info = Pa_GetDeviceInfo(input_device_);
-    const PaDeviceInfo* out_info = Pa_GetDeviceInfo(output_device_);
-    std::cout << "Audio stream started:" << std::endl;
-    std::cout << "  Input:   " << (in_info ? in_info->name : "Unknown") << std::endl;
-    std::cout << "  Output:  " << (out_info ? out_info->name : "Unknown") << std::endl;
-    std::cout << "  Rate:    " << (si ? si->sampleRate : sample_rate_) << " Hz" << std::endl;
-    if (si) {
-        std::cout << "  Latency: in=" << (si->inputLatency * 1000.0) << " ms"
-                  << "  out=" << (si->outputLatency * 1000.0) << " ms" << std::endl;
-    }
-    return true;
-}
-
-void AudioEngine::stop() {
-    if (stream_) {
-        if (running_) {
-            Pa_StopStream(stream_);
-            running_ = false;
-        }
-        Pa_CloseStream(stream_);
-        stream_ = nullptr;
-    }
-}
-
-bool AudioEngine::restart() {
-    stop();
-    bool ok = start();
-    if (!ok) {
-        last_error_ = "Failed to restart audio stream. Check device settings.";
-        std::cerr << "[Amplitron] " << last_error_ << std::endl;
-    } else {
-        last_error_.clear();
-    }
-    return ok;
-}
-
-bool AudioEngine::devices_share_host_api(int input_dev, int output_dev) const {
-    const PaDeviceInfo* in_info = Pa_GetDeviceInfo(input_dev);
-    const PaDeviceInfo* out_info = Pa_GetDeviceInfo(output_dev);
-    if (!in_info || !out_info) return false;
-    return in_info->hostApi == out_info->hostApi;
-}
-
-std::string AudioEngine::get_input_device_name() const {
-    if (input_device_ >= 0) {
-        const PaDeviceInfo* info = Pa_GetDeviceInfo(input_device_);
-        if (info) return info->name;
-    }
-    return "None";
-}
-
-std::string AudioEngine::get_output_device_name() const {
-    if (output_device_ >= 0) {
-        const PaDeviceInfo* info = Pa_GetDeviceInfo(output_device_);
-        if (info) return info->name;
-    }
-    return "None";
-}
-
-std::vector<AudioDeviceInfo> AudioEngine::get_input_devices() const {
-    std::vector<AudioDeviceInfo> devices;
-    int count = Pa_GetDeviceCount();
-    for (int i = 0; i < count; ++i) {
-        const PaDeviceInfo* info = Pa_GetDeviceInfo(i);
-        if (info && info->maxInputChannels > 0) {
-            devices.push_back({
-                i, info->name,
-                info->maxInputChannels, info->maxOutputChannels,
-                info->defaultSampleRate,
-                is_usb_device_name(info->name)
-            });
-        }
-    }
-    return devices;
-}
-
-std::vector<AudioDeviceInfo> AudioEngine::get_output_devices() const {
-    std::vector<AudioDeviceInfo> devices;
-    int count = Pa_GetDeviceCount();
-    for (int i = 0; i < count; ++i) {
-        const PaDeviceInfo* info = Pa_GetDeviceInfo(i);
-        if (info && info->maxOutputChannels > 0) {
-            devices.push_back({
-                i, info->name,
-                info->maxInputChannels, info->maxOutputChannels,
-                info->defaultSampleRate,
-                is_usb_device_name(info->name)
-            });
-        }
-    }
-    return devices;
-}
-
-bool AudioEngine::set_input_device(int device_index) {
-    const PaDeviceInfo* info = Pa_GetDeviceInfo(device_index);
-    if (!info || info->maxInputChannels < 1) {
-        last_error_ = "Invalid input device.";
-        return false;
-    }
-
-    // Warn about host API mismatch (PortAudio requires same API)
-    if (!devices_share_host_api(device_index, output_device_)) {
-        const PaDeviceInfo* out_info = Pa_GetDeviceInfo(output_device_);
-        const PaHostApiInfo* in_api = Pa_GetHostApiInfo(info->hostApi);
-        const PaHostApiInfo* out_api = out_info ? Pa_GetHostApiInfo(out_info->hostApi) : nullptr;
-        std::cerr << "[Amplitron] Warning: Input (" << (in_api ? in_api->name : "?") 
-                  << ") and output (" << (out_api ? out_api->name : "?") 
-                  << ") are on different host APIs. Stream may fail." << std::endl;
-    }
-
-    int prev_device = input_device_;
-    bool was_running = running_;
-    if (was_running) stop();
-    input_device_ = device_index;
-    if (was_running) {
-        if (!start()) {
-            last_error_ = "Failed to start with new input device. Reverting.";
-            std::cerr << "[Amplitron] " << last_error_ << std::endl;
-            input_device_ = prev_device;
-            start();  // restore previous
-            return false;
-        }
-        last_error_.clear();
-    }
-    return true;
-}
-
-bool AudioEngine::set_output_device(int device_index) {
-    const PaDeviceInfo* info = Pa_GetDeviceInfo(device_index);
-    if (!info || info->maxOutputChannels < 1) {
-        last_error_ = "Invalid output device.";
-        return false;
-    }
-
-    if (!devices_share_host_api(input_device_, device_index)) {
-        const PaDeviceInfo* in_info = Pa_GetDeviceInfo(input_device_);
-        const PaHostApiInfo* in_api = in_info ? Pa_GetHostApiInfo(in_info->hostApi) : nullptr;
-        const PaHostApiInfo* out_api = Pa_GetHostApiInfo(info->hostApi);
-        std::cerr << "[Amplitron] Warning: Input (" << (in_api ? in_api->name : "?") 
-                  << ") and output (" << (out_api ? out_api->name : "?") 
-                  << ") are on different host APIs. Stream may fail." << std::endl;
-    }
-
-    int prev_device = output_device_;
-    bool was_running = running_;
-    if (was_running) stop();
-    output_device_ = device_index;
-    if (was_running) {
-        if (!start()) {
-            last_error_ = "Failed to start with new output device. Reverting.";
-            std::cerr << "[Amplitron] " << last_error_ << std::endl;
-            output_device_ = prev_device;
-            start();
-            return false;
-        }
-        last_error_.clear();
-    }
-    return true;
-}
-
-#else // __EMSCRIPTEN__
-// --- Web / Emscripten SDL_Audio implementation ---
-
-void AudioEngine::sdl_audio_callback(void* userdata, Uint8* stream, int len) {
-    auto* engine = static_cast<AudioEngine*>(userdata);
-    auto* out = reinterpret_cast<float*>(stream);
-    int frame_count = len / static_cast<int>(sizeof(float));
-
-    auto& cap = engine->capture_buffer_;
-    if (static_cast<int>(cap.size()) < frame_count)
-        cap.resize(static_cast<size_t>(frame_count), 0.0f);
-
-    if (engine->sdl_capture_device_ != 0) {
-        Uint32 queued = SDL_GetQueuedAudioSize(engine->sdl_capture_device_);
-        Uint32 need = static_cast<Uint32>(frame_count * sizeof(float));
-
-        // Drain excess to prevent latency buildup (keep at most 2 buffers ahead)
-        Uint32 max_queued = need * 2;
-        while (queued > max_queued) {
-            // Discard in small chunks to avoid any allocation
-            Uint8 junk[4096];
-            Uint32 chunk = (queued - need) > 4096 ? 4096 : (queued - need);
-            SDL_DequeueAudio(engine->sdl_capture_device_, junk, chunk);
-            queued -= chunk;
-        }
-
-        Uint32 got = SDL_DequeueAudio(engine->sdl_capture_device_,
-                                      cap.data(), need);
-        int captured = static_cast<int>(got / sizeof(float));
-        if (captured < frame_count)
-            std::memset(cap.data() + captured, 0,
-                        static_cast<size_t>(frame_count - captured) * sizeof(float));
-    } else {
-        std::memset(cap.data(), 0, static_cast<size_t>(frame_count) * sizeof(float));
-    }
-
-    engine->process_audio(cap.data(), out, frame_count);
-}
-
-bool AudioEngine::initialize() {
-    if (SDL_InitSubSystem(SDL_INIT_AUDIO) != 0) {
-        std::cerr << "SDL audio init failed: " << SDL_GetError() << std::endl;
-        return false;
-    }
-    initialized_ = true;
-    std::cout << "[Web] Audio subsystem initialized." << std::endl;
-    return true;
-}
-
-void AudioEngine::shutdown() {
-    stop();
-    if (initialized_) {
-        SDL_QuitSubSystem(SDL_INIT_AUDIO);
-        initialized_ = false;
-    }
-}
-
-bool AudioEngine::start() {
-    if (!initialized_ || running_) return false;
-
-    // Web Audio needs buffer 256–16384; use 1024 for glitch-free playback (~21ms at 48kHz)
-    int web_buffer = 1024;
-
-    // --- Open output (playback) device with callback ---
-    SDL_AudioSpec want_out, have_out;
-    SDL_memset(&want_out, 0, sizeof(want_out));
-    want_out.freq = sample_rate_;
-    want_out.format = AUDIO_F32;
-    want_out.channels = 1;
-    want_out.samples = static_cast<Uint16>(web_buffer);
-    want_out.callback = sdl_audio_callback;
-    want_out.userdata = this;
-
-    // Allow frequency change so browser uses its native sample rate (no resampling artifacts)
-    sdl_audio_device_ = SDL_OpenAudioDevice(nullptr, 0, &want_out, &have_out,
-                                            SDL_AUDIO_ALLOW_FREQUENCY_CHANGE);
-    if (sdl_audio_device_ == 0) {
-        std::cerr << "[Web] Failed to open output audio: " << SDL_GetError() << std::endl;
-        return false;
-    }
-
-    sample_rate_ = have_out.freq;
-    buffer_size_ = have_out.samples;
-
-    // Update all effects with the actual sample rate the browser gave us
-    {
-        std::lock_guard<std::mutex> lock(effect_mutex_);
-        for (auto& fx : effects_)
-            fx->set_sample_rate(sample_rate_);
-    }
-
-    // --- Open capture (microphone) device ---
-    // Enumerate capture devices and prefer USB/external audio interfaces
-    SDL_AudioSpec want_cap, have_cap;
-    SDL_memset(&want_cap, 0, sizeof(want_cap));
-    want_cap.freq = sample_rate_;
-    want_cap.format = AUDIO_F32;
-    want_cap.channels = 1;
-    want_cap.samples = static_cast<Uint16>(web_buffer);
-    want_cap.callback = nullptr;  // use SDL_DequeueAudio for capture
-
-    // Keywords matching USB guitar cables / audio interfaces
-    static const char* usb_keywords[] = {
-        "usb", "guitar", "irig", "scarlett", "behringer", "focusrite",
-        "presonus", "steinberg", "audio interface", "line 6", "rocksmith",
-        "umc", "um2", "uphoria", "podcast", "xenyx", "external"
-    };
-
-    const char* preferred_device = nullptr;
-    int num_capture = SDL_GetNumAudioDevices(1);
-    std::cout << "[Web] Found " << num_capture << " capture device(s):" << std::endl;
-    for (int i = 0; i < num_capture; ++i) {
-        const char* name = SDL_GetAudioDeviceName(i, 1);
-        if (!name) continue;
-        std::cout << "  [" << i << "] " << name << std::endl;
-
-        // Check if this looks like an external audio interface
-        std::string lower = name;
-        std::transform(lower.begin(), lower.end(), lower.begin(),
-                       [](unsigned char c) { return std::tolower(c); });
-        for (const auto* kw : usb_keywords) {
-            if (lower.find(kw) != std::string::npos) {
-                preferred_device = name;
-                std::cout << "  >> Preferred (matches '" << kw << "')" << std::endl;
-                break;
-            }
-        }
-        if (preferred_device) break;
-    }
-
-    sdl_capture_device_ = SDL_OpenAudioDevice(preferred_device, 1, &want_cap, &have_cap,
-                                                SDL_AUDIO_ALLOW_FREQUENCY_CHANGE);
-    if (sdl_capture_device_ == 0) {
-        std::cerr << "[Web] Warning: No microphone available: " << SDL_GetError() << std::endl;
-        // Continue without mic — effects still run on silence
-    }
-
-    // Pre-allocate capture buffer
-    capture_buffer_.resize(static_cast<size_t>(buffer_size_), 0.0f);
-
-    // Unpause both devices
-    SDL_PauseAudioDevice(sdl_audio_device_, 0);
-    if (sdl_capture_device_ != 0)
-        SDL_PauseAudioDevice(sdl_capture_device_, 0);
-
-    running_ = true;
-
-    std::cout << "[Web] Audio stream started:" << std::endl;
-    std::cout << "  Rate:   " << sample_rate_ << " Hz" << std::endl;
-    std::cout << "  Buffer: " << buffer_size_ << " samples" << std::endl;
-    std::cout << "  Mic:    " << (sdl_capture_device_ ? "yes" : "unavailable") << std::endl;
-    return true;
-}
-
-void AudioEngine::stop() {
-    if (running_) {
-        if (sdl_audio_device_ != 0)
-            SDL_PauseAudioDevice(sdl_audio_device_, 1);
-        if (sdl_capture_device_ != 0)
-            SDL_PauseAudioDevice(sdl_capture_device_, 1);
-        running_ = false;
-    }
-    if (sdl_capture_device_ != 0) {
-        SDL_CloseAudioDevice(sdl_capture_device_);
-        sdl_capture_device_ = 0;
-    }
-    if (sdl_audio_device_ != 0) {
-        SDL_CloseAudioDevice(sdl_audio_device_);
-        sdl_audio_device_ = 0;
-    }
-}
-
-bool AudioEngine::restart() {
-    stop();
-    bool ok = start();
-    if (!ok) {
-        last_error_ = "Failed to restart audio.";
-    } else {
-        last_error_.clear();
-    }
-    return ok;
-}
-
-// Web stubs for device management (browser handles devices)
-std::string AudioEngine::get_input_device_name() const { return "Browser Microphone"; }
-std::string AudioEngine::get_output_device_name() const { return "Browser Audio Output"; }
-std::vector<AudioDeviceInfo> AudioEngine::get_input_devices() const {
-    return {{0, "Browser Microphone", 1, 0, 48000.0, false}};
-}
-std::vector<AudioDeviceInfo> AudioEngine::get_output_devices() const {
-    return {{0, "Browser Audio Output", 0, 1, 48000.0, false}};
-}
-bool AudioEngine::set_input_device(int) { return true; }
-bool AudioEngine::set_output_device(int) { return true; }
-
-#endif // __EMSCRIPTEN__
 
 void AudioEngine::add_effect(std::shared_ptr<Effect> effect) {
     std::lock_guard<std::mutex> lock(effect_mutex_);
@@ -727,6 +39,15 @@ void AudioEngine::remove_effect(int index) {
     }
 }
 
+void AudioEngine::restore_effects_state(std::vector<std::shared_ptr<Effect>> new_effects) {
+    std::lock_guard<std::mutex> lock(effect_mutex_);
+    effects_.clear();
+    for (auto& fx : new_effects) {
+        fx->set_sample_rate(sample_rate_);
+        effects_.push_back(std::move(fx));
+    }
+}
+
 void AudioEngine::move_effect(int from, int to) {
     std::lock_guard<std::mutex> lock(effect_mutex_);
     int n = static_cast<int>(effects_.size());
@@ -735,6 +56,10 @@ void AudioEngine::move_effect(int from, int to) {
     effects_.erase(effects_.begin() + from);
     effects_.insert(effects_.begin() + to, effect);
 }
+
+// =============================================================================
+// Buffer / sample rate settings (call stop/start which are backend-provided)
+// =============================================================================
 
 void AudioEngine::set_buffer_size(int size) {
     size = std::max(MIN_BUFFER_SIZE, std::min(MAX_BUFFER_SIZE, size));
@@ -783,25 +108,9 @@ void AudioEngine::set_sample_rate(int rate) {
     }
 }
 
-#ifndef __EMSCRIPTEN__
-int AudioEngine::audio_callback(const void* input, void* output,
-                                 unsigned long frame_count,
-                                 const PaStreamCallbackTimeInfo* /*time_info*/,
-                                 PaStreamCallbackFlags /*status_flags*/,
-                                 void* user_data) {
-    auto* engine = static_cast<AudioEngine*>(user_data);
-    const auto* in = static_cast<const float*>(input);
-    auto* out = static_cast<float*>(output);
-
-    if (!in || !out) {
-        if (out) std::memset(out, 0, frame_count * sizeof(float));
-        return paContinue;
-    }
-
-    engine->process_audio(in, out, static_cast<int>(frame_count));
-    return paContinue;
-}
-#endif
+// =============================================================================
+// Audio processing pipeline (called from platform callback)
+// =============================================================================
 
 void AudioEngine::process_audio(const float* input, float* output, int frame_count) {
     auto t_start = std::chrono::steady_clock::now();
@@ -860,7 +169,9 @@ void AudioEngine::process_audio(const float* input, float* output, int frame_cou
     cpu_load_.store(duration_us / budget_us, std::memory_order_relaxed);
 }
 
-// --- Lock-free command drain (called from audio thread) ---
+// =============================================================================
+// Lock-free command drain (called from audio thread)
+// =============================================================================
 
 void AudioEngine::drain_commands() {
     AudioCommand cmd;
@@ -900,7 +211,9 @@ void AudioEngine::drain_commands() {
     }
 }
 
-// --- GUI-thread push methods ---
+// =============================================================================
+// GUI-thread SPSC push methods
+// =============================================================================
 
 void AudioEngine::set_input_gain(float gain) {
     AudioCommand cmd{};
@@ -943,7 +256,9 @@ void AudioEngine::push_effect_mix(int effect_index, float mix) {
     command_queue_.try_push(cmd);
 }
 
-// --- Buffer auto-tuning ---
+// =============================================================================
+// Buffer auto-tuning
+// =============================================================================
 
 int AudioEngine::get_suggested_buffer_size() const {
     float load = cpu_load_.load(std::memory_order_relaxed);
