+++ src/gui/command.h
@@ -3,43 +3,72 @@
 #include "common.h"
 #include "audio/audio_engine.h"
 #include "audio/effect.h"
+#include <algorithm>
 #include <chrono>
+#include <memory>
+#include <utility>
+#include <vector>
 
 namespace GuitarAmp {
 
-// Base class for all undoable commands (Gang of Four Command Pattern)
+/**
+ * @brief Abstract base class for all undoable commands (Gang of Four Command Pattern).
+ *
+ * Each concrete command encapsulates a single reversible action on the audio
+ * engine (e.g. adding an effect, changing a parameter). Commands are stored
+ * in a CommandHistory and invoked via execute() / undo().
+ */
 class Command {
 public:
     virtual ~Command() = default;
+
+    /** @brief Apply this command's action. */
     virtual void execute() = 0;
+
+    /** @brief Reverse this command's action. */
     virtual void undo() = 0;
+
+    /** @brief Return a short human-readable label (shown in the Edit menu). */
     virtual const char* description() const = 0;
 
-    // For coalescing: two commands can merge if they affect the same target
-    // within a short time window. Returns true if this command absorbed `other`.
+    /**
+     * @brief Attempt to merge @p other into this command (coalescing).
+     *
+     * Two commands can merge if they affect the same target within a short
+     * time window. Returns true if this command absorbed @p other.
+     */
     virtual bool merge_with(const Command& /*other*/) { return false; }
 
+    /** @brief Return the steady-clock time point when this command was created. */
     auto timestamp() const { return timestamp_; }
 
 protected:
     std::chrono::steady_clock::time_point timestamp_ = std::chrono::steady_clock::now();
 };
 
-// ---------------------------------------------------------------------------
-// AddEffectCommand
-// ---------------------------------------------------------------------------
+/**
+ * @brief Command that appends an effect to the engine's signal chain.
+ *
+ * execute() adds the effect; undo() finds and removes it.
+ */
 class AddEffectCommand : public Command {
 public:
+    /**
+     * @brief Construct an AddEffectCommand.
+     * @param engine  Reference to the audio engine that owns the effect chain.
+     * @param effect  Shared pointer to the effect to add.
+     */
     AddEffectCommand(AudioEngine& engine, std::shared_ptr<Effect> effect)
         : engine_(engine), effect_(std::move(effect)) {}
 
+    /** @brief Append the effect to the engine's chain. */
     void execute() override {
         engine_.add_effect(effect_);
     }
 
+    /** @brief Remove the previously added effect from the chain. */
     void undo() override {
         auto& fx = engine_.effects();
-        // Find and remove the effect we added (should be at the end)
         for (int i = static_cast<int>(fx.size()) - 1; i >= 0; --i) {
             if (fx[i] == effect_) {
                 engine_.remove_effect(i);
@@ -48,8 +77,10 @@ public:
         }
     }
 
+    /** @brief Return "Add Effect". */
     const char* description() const override { return "Add Effect"; }
 
+    /** @brief Accessor for the wrapped effect. */
     std::shared_ptr<Effect> effect() const { return effect_; }
 
 private:
@@ -57,31 +88,38 @@ private:
     std::shared_ptr<Effect> effect_;
 };
 
-// ---------------------------------------------------------------------------
-// RemoveEffectCommand
-// ---------------------------------------------------------------------------
+/**
+ * @brief Command that removes an effect from the engine's signal chain.
+ *
+ * The constructor captures the effect pointer before removal so undo() can
+ * re-insert it at its original position.
+ */
 class RemoveEffectCommand : public Command {
 public:
+    /**
+     * @brief Construct a RemoveEffectCommand.
+     * @param engine  Reference to the audio engine.
+     * @param index   Chain index of the effect to remove (captured at construction).
+     */
     RemoveEffectCommand(AudioEngine& engine, int index)
         : engine_(engine), index_(index) {
-        // Capture the effect before removal
         auto& fx = engine_.effects();
         if (index >= 0 && index < static_cast<int>(fx.size())) {
             effect_ = fx[index];
         }
     }
 
+    /** @brief Remove the effect at the stored index. */
     void execute() override {
         engine_.remove_effect(index_);
     }
 
+    /** @brief Re-insert the captured effect at its original chain position. */
     void undo() override {
         if (effect_) {
-            // Re-insert at original position
             auto& fx = engine_.effects();
             int pos = std::min(index_, static_cast<int>(fx.size()));
             engine_.add_effect(effect_);
-            // add_effect appends; move it to the correct position
             int last = static_cast<int>(engine_.effects().size()) - 1;
             if (last != pos) {
                 engine_.move_effect(last, pos);
@@ -89,9 +127,13 @@ public:
         }
     }
 
+    /** @brief Return "Remove Effect". */
     const char* description() const override { return "Remove Effect"; }
 
+    /** @brief Original chain index of the removed effect. */
     int index() const { return index_; }
+
+    /** @brief Accessor for the captured effect pointer. */
     std::shared_ptr<Effect> effect() const { return effect_; }
 
 private:
@@ -100,25 +142,39 @@ private:
     std::shared_ptr<Effect> effect_;
 };
 
-// ---------------------------------------------------------------------------
-// ReorderEffectCommand
-// ---------------------------------------------------------------------------
+/**
+ * @brief Command that moves an effect from one chain position to another.
+ *
+ * execute() calls move_effect(from, to); undo() reverses the move.
+ */
 class ReorderEffectCommand : public Command {
 public:
+    /**
+     * @brief Construct a ReorderEffectCommand.
+     * @param engine  Reference to the audio engine.
+     * @param from    Source index in the effect chain.
+     * @param to      Destination index in the effect chain.
+     */
     ReorderEffectCommand(AudioEngine& engine, int from, int to)
         : engine_(engine), from_(from), to_(to) {}
 
+    /** @brief Move the effect from source to destination index. */
     void execute() override {
         engine_.move_effect(from_, to_);
     }
 
+    /** @brief Move the effect back from destination to source index. */
     void undo() override {
         engine_.move_effect(to_, from_);
     }
 
+    /** @brief Return "Reorder Effect". */
     const char* description() const override { return "Reorder Effect"; }
 
+    /** @brief Original source index. */
     int from() const { return from_; }
+
+    /** @brief Destination index. */
     int to() const { return to_; }
 
 private:
@@ -127,17 +183,28 @@ private:
     int to_;
 };
 
-// ---------------------------------------------------------------------------
-// ParameterChangeCommand (supports coalescing)
-// Uses shared_ptr<Effect> directly — robust against effect reordering.
-// ---------------------------------------------------------------------------
+/**
+ * @brief Command that records a single parameter value change on an effect.
+ *
+ * Supports coalescing: rapid changes to the same parameter within 500 ms are
+ * merged into one undo step. Uses shared_ptr<Effect> directly so it is robust
+ * against effect chain reordering.
+ */
 class ParameterChangeCommand : public Command {
 public:
+    /**
+     * @brief Construct a ParameterChangeCommand.
+     * @param effect      Shared pointer to the target effect.
+     * @param param_index Index of the parameter within the effect's param list.
+     * @param old_value   Value before the change (used by undo).
+     * @param new_value   Value after the change (used by execute).
+     */
     ParameterChangeCommand(std::shared_ptr<Effect> effect,
                            int param_index, float old_value, float new_value)
         : effect_(std::move(effect)),
           param_index_(param_index), old_value_(old_value), new_value_(new_value) {}
 
+    /** @brief Set the parameter to new_value_. */
     void execute() override {
         auto& params = effect_->params();
         if (param_index_ >= 0 && param_index_ < static_cast<int>(params.size())) {
@@ -145,6 +212,7 @@ public:
         }
     }
 
+    /** @brief Restore the parameter to old_value_. */
     void undo() override {
         auto& params = effect_->params();
         if (param_index_ >= 0 && param_index_ < static_cast<int>(params.size())) {
@@ -152,28 +220,43 @@ public:
         }
     }
 
+    /** @brief Return "Change Parameter". */
     const char* description() const override { return "Change Parameter"; }
 
+    /**
+     * @brief Attempt to coalesce @p other into this command.
+     *
+     * Merges if @p other is a ParameterChangeCommand targeting the same
+     * effect and parameter index within 500 ms. On success, this command's
+     * new_value_ and timestamp are updated; old_value_ is preserved.
+     *
+     * @return true if @p other was absorbed.
+     */
     bool merge_with(const Command& other) override {
         auto* pc = dynamic_cast<const ParameterChangeCommand*>(&other);
         if (!pc) return false;
         if (pc->effect_.get() != effect_.get() || pc->param_index_ != param_index_)
             return false;
 
-        // Coalesce if within 500ms
         auto dt = std::chrono::duration_cast<std::chrono::milliseconds>(
             pc->timestamp_ - timestamp_);
         if (dt.count() > 500) return false;
 
-        // Absorb: keep our old_value_, update new_value_ and timestamp
         new_value_ = pc->new_value_;
         timestamp_ = pc->timestamp_;
         return true;
     }
 
+    /** @brief Accessor for the target effect. */
     std::shared_ptr<Effect> effect() const { return effect_; }
+
+    /** @brief Index of the parameter being changed. */
     int param_index() const { return param_index_; }
+
+    /** @brief Value before the change. */
     float old_value() const { return old_value_; }
+
+    /** @brief Value after the change. */
     float new_value() const { return new_value_; }
 
 private:
@@ -183,20 +266,35 @@ private:
     float new_value_;
 };
 
-// ---------------------------------------------------------------------------
-// LoadPresetCommand
-// ---------------------------------------------------------------------------
+/**
+ * @brief Command that captures the full effect-chain state before and after a
+ *        preset load, enabling undo/redo of the entire preset switch.
+ *
+ * Uses AudioEngine::restore_effects_state() so the effect chain is replaced
+ * atomically under the engine's mutex.
+ */
 class LoadPresetCommand : public Command {
 public:
-    // Captures the full state before and after a preset load.
-    // before_effects / after_effects hold shared_ptrs + param snapshots.
+    /**
+     * @brief Snapshot of a single effect's configuration at a point in time.
+     */
     struct EffectSnapshot {
-        std::shared_ptr<Effect> effect;
-        bool enabled;
-        float mix;
-        std::vector<float> param_values;
+        std::shared_ptr<Effect> effect;   ///< The effect instance.
+        bool enabled;                     ///< Whether the effect was enabled.
+        float mix;                        ///< Dry/wet mix level.
+        std::vector<float> param_values;  ///< Ordered parameter values.
     };
 
+    /**
+     * @brief Construct a LoadPresetCommand.
+     * @param engine             Reference to the audio engine.
+     * @param before_state       Effect chain snapshot before the load.
+     * @param before_input_gain  Input gain before the load.
+     * @param before_output_gain Output gain before the load.
+     * @param after_state        Effect chain snapshot after the load.
+     * @param after_input_gain   Input gain after the load.
+     * @param after_output_gain  Output gain after the load.
+     */
     LoadPresetCommand(AudioEngine& engine,
                       std::vector<EffectSnapshot> before_state,
                       float before_input_gain, float before_output_gain,
@@ -210,26 +308,31 @@ public:
           after_input_gain_(after_input_gain),
           after_output_gain_(after_output_gain) {}
 
+    /** @brief Restore the after-load state (redo). */
     void execute() override {
         apply_state(after_state_, after_input_gain_, after_output_gain_);
     }
 
+    /** @brief Restore the before-load state (undo). */
     void undo() override {
         apply_state(before_state_, before_input_gain_, before_output_gain_);
     }
 
+    /** @brief Return "Load Preset". */
     const char* description() const override { return "Load Preset"; }
 
 private:
+    /**
+     * @brief Replace the engine's effect chain and gains with the given snapshot.
+     * @param state        Vector of EffectSnapshot to restore.
+     * @param input_gain   Input gain to set.
+     * @param output_gain  Output gain to set.
+     */
     void apply_state(const std::vector<EffectSnapshot>& state,
                      float input_gain, float output_gain) {
-        // Clear current effects
-        auto& fx = engine_.effects();
-        while (!fx.empty()) {
-            engine_.remove_effect(static_cast<int>(fx.size()) - 1);
-        }
-
-        // Restore effects
+        // Prepare effect list with restored params before swapping
+        std::vector<std::shared_ptr<Effect>> new_effects;
+        new_effects.reserve(state.size());
         for (auto& snap : state) {
             snap.effect->set_enabled(snap.enabled);
             snap.effect->set_mix(snap.mix);
@@ -238,8 +341,11 @@ private:
                             i < static_cast<int>(snap.param_values.size()); ++i) {
                 params[i].value = snap.param_values[i];
             }
-            engine_.add_effect(snap.effect);
+            new_effects.push_back(snap.effect);
         }
+
+        // Atomic swap under engine lock — audio thread never sees half-applied state
+        engine_.restore_effects_state(std::move(new_effects));
 
         engine_.set_input_gain(input_gain);
         engine_.set_output_gain(output_gain);
