+++ src/gui/command_history.cpp
@@ -2,8 +2,8 @@
 
 namespace GuitarAmp {
 
+/** @brief Execute cmd, attempt coalescing, push to undo stack, clear redo. */
 void CommandHistory::execute(std::unique_ptr<Command> cmd) {
-    // Execute the command first
     cmd->execute();
 
     // Try coalescing with the top of the undo stack
@@ -18,8 +18,8 @@ void CommandHistory::execute(std::unique_ptr<Command> cmd) {
     redo_stack_.clear();
 }
 
+/** @brief Push an already-applied command; attempt coalescing, clear redo. */
 void CommandHistory::push_executed(std::unique_ptr<Command> cmd) {
-    // Try coalescing with the top of the undo stack
     if (!undo_stack_.empty() && undo_stack_.back()->merge_with(*cmd)) {
         // Merged into existing top — no new entry needed
     } else {
@@ -31,6 +31,7 @@ void CommandHistory::push_executed(std::unique_ptr<Command> cmd) {
     redo_stack_.clear();
 }
 
+/** @brief Pop the top undo command, call its undo(), and move it to redo. */
 bool CommandHistory::undo() {
     if (undo_stack_.empty()) return false;
 
@@ -42,6 +43,7 @@ bool CommandHistory::undo() {
     return true;
 }
 
+/** @brief Pop the top redo command, call its execute(), and move it to undo. */
 bool CommandHistory::redo() {
     if (redo_stack_.empty()) return false;
 
@@ -53,21 +55,25 @@ bool CommandHistory::redo() {
     return true;
 }
 
+/** @brief Discard all undo and redo history. */
 void CommandHistory::clear() {
     undo_stack_.clear();
     redo_stack_.clear();
 }
 
+/** @brief Return the description of the top undo command, or nullptr. */
 const char* CommandHistory::undo_description() const {
     if (undo_stack_.empty()) return nullptr;
     return undo_stack_.back()->description();
 }
 
+/** @brief Return the description of the top redo command, or nullptr. */
 const char* CommandHistory::redo_description() const {
     if (redo_stack_.empty()) return nullptr;
     return redo_stack_.back()->description();
 }
 
+/** @brief Remove oldest undo entries until stack size <= max_depth_. */
 void CommandHistory::trim() {
     while (static_cast<int>(undo_stack_.size()) > max_depth_) {
         undo_stack_.erase(undo_stack_.begin());
